Fix bug in req func, tips down func can work now again.

+++ utility/backup.py
@@ -57,7 +57,8 @@ def backup_t0002(src_path, en_compress=True):
 def recover_t0002(dst_path, en_compress=True):
     src_folder_name = 't0002_backup'
     src_path = os.path.join(src_folder_name, '*.*')
-    shutil.rmtree(os.path.join(dst_path, 'pad'))
+    if os.path.exists(os.path.join(dst_path, 'pad')):
+        shutil.rmtree(os.path.join(dst_path, 'pad'))
 
     if en_compress:
         cmd = '{} x -y {}.7z -o{}'.format(sets.EXE_7Z, src_folder_name, dst_path)
